refactor(invoice): split the HTML generator out of the invoice template
src/utils/invoice/template.ts was 1191 lines, over the 750-line ceiling.
generateInvoiceHTML alone accounts for 715 of them.

Move it verbatim into src/utils/invoice/html-template.ts and re-export it
from template.ts, so no caller changes.

The three asset unit constants and formatCryptoUnitConversion move into
src/utils/invoice/crypto-units.ts. Both template.ts and html-template.ts
read them from there. Importing them from template.ts instead would make
html-template.ts and template.ts import each other at run time.
template.ts re-exports all four, so pdf-generator.ts keeps working.

One whitespace-only line inside a CSS block in the template literal loses
its trailing indentation to prettier. The generated HTML is not hashed,
and the line carries no markup, so the rendered invoice is unchanged.

Ported from #738, which is superseded.
